requirements.txt
streamlit
poker
langchain
openai
python-dotenv
langchain-community